requirements.txt
line-bot-sdk==1.17.0
loguru==0.5.3
requests==2.24.0
